fix(perfs): remove duplicate API calls in new_explorer=1
